backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
PyMuPDF>=1.24.0
python-docx>=1.1.0
spacy>=3.7.0
keybert>=0.8.0
sentence-transformers>=3.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pydantic>=2.8.0
aiofiles>=23.2.0
